ShopmeWebParent/ShopmeBackend/src/main/java/com/shopme/admin/brand/BrandService.java
package com.shopme.admin.brand;

import com.shopme.common.entity.Brand;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.NoSuchElementException;
import java.util.Objects;

@Service
public class BrandService {

    @Autowired
    private BrandRepository brandRepository;

    public List<Brand> listAll() {
        return (List<Brand>) brandRepository.findAll();
    }

    public Brand save(Brand brand) {
        return brandRepository.save(brand);
    }

    public Brand get(Integer id) throws BrandNotFoundException {
        try {
            return brandRepository.findById(id).get();
        } catch (NoSuchElementException exception) {
            throw new BrandNotFoundException("Could not find any brand with ID " + id);
        }
    }

    public void delete(Integer id) throws BrandNotFoundException {
        Long countById = brandRepository.countById(id);

        if (null == countById || 0 == countById) {
            throw new BrandNotFoundException("Could not find any brand with ID " + id);
        }

        brandRepository.deleteById(id);
    }

    public String checkUnique(Integer id, String name) {
        boolean isCreatingNew = (null == id || 0 == id);
        Brand brandByName = brandRepository.findByName(name);

        if (isCreatingNew) {
            if (null != brandByName) {
                return "Duplicate";
            }
        } else {
            if (null != brandByName && !Objects.equals(id, brandByName.getId())) {
                return "Duplicate";
            }
        }

        return "OK";
    }
}
